requirements.txt
jinja2>=3,<4
click>=8,<9
pyyaml>=6,<7